pyproject.toml
[project]
name = "ai-job-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
